+++ test/event_test.dart
@@ -76,56 +76,98 @@ void main() {
       expect(e.subscriberCount, equals(0));
     });
 
-    test('Calling broadcast successful with no args handler', () {
+    test('Calling broadcast is successful with no args handler', () {
       bool changedInHandler = false;
 
       var e = Event();
-      e.subscribe((args) => {changedInHandler = true});
+      e.subscribe((args) => changedInHandler = true);
       e.broadcast();
       expect(changedInHandler, equals(true));
     });
 
-    test('Calling broadcast with argument successful', () {
+    test('Calling broadcast with typed Value successful', () {
       // a value expected to change when event is raised
-      int? changedInHandler = -1;
+      int changedInHandler = -1;
 
       var e = Event<Value<int>>();
-      e.subscribe((args) => {changedInHandler = args?.value});
+      e.subscribe((args) => changedInHandler = args.value);
       e.broadcast(Value(39));
 
       expect(changedInHandler, equals(39));
     });
 
-    test('Use WhenWhy EventArg fields (whenOccurred and description) correctly', () {
-      DateTime? eventOccurred;
-      String? description;
+    test('Calling broadcast with untyped Value is successful', () {
+      // a value expected to change when event is raised
+      int changedInHandler = -1;
 
-      var e = Event<WhenWhy>();
+      var e = Event<Value>();
+      e.subscribe((args) => changedInHandler = args.value);
+      e.broadcast(Value(41));
+
+      expect(changedInHandler, equals(41));
+    });
+
+    test('Use whenOccured EventArg field correctly', () {
+      DateTime? eventOccurred;
+
+      var e = Event(); // <EventArgs> assumed
       e.subscribe((args) {
-        eventOccurred = args?.whenOccurred;
-        description = args?.description;
+        eventOccurred = args.whenOccurred;
       });
-      e.broadcast(WhenWhy(description: 'test description'));
+      e.broadcast();
       expect(DateTime.now().difference(eventOccurred!), lessThan(Duration(milliseconds: 250)));
-      expect(description, equals('test description'));
+    });
+
+    test('Can set an event name which is then available in args', () {
+      String? eventName;
+
+      var e = Event('myTestEvent');
+      e.subscribe((args) {
+        eventName = args.eventName;
+      });
+      e.broadcast();
+
+      expect(eventName, "myTestEvent");
+    });
+
+    test('Setting no event name results in the arg name value of ""', () {
+      String? eventName;
+
+      var e = Event();
+      e.subscribe((args) {
+        eventName = args.eventName;
+      });
+      e.broadcast();
+
+      expect(eventName, "");
     });
 
     test('Single Value as a generic Event argument', () {
       String? value;
 
       var e = Event<Value<String>>();
-      e.subscribe((args) => value = args?.value);
+      e.subscribe((args) => value = args.value);
       e.broadcast(Value('hello'));
+
       expect(value, equals('hello'));
     });
 
     test('Single Value as a generic argument, where value type is dynamic', () {
-      String? value;
+      String value = '';
 
       var e = Event<Value>(); // Value is implicitly a string per the constructor
-      e.subscribe((args) => value = args?.value);
+      e.subscribe((args) => value = args.value);
       e.broadcast(Value('any string'));
       expect(value, equals('any string'));
+    });
+
+    test('Using Value as a Map', () {
+      Map map = {};
+
+      var e = Event<Value>(); // Value is implicitly a Map per the constructor
+      e.subscribe((args) => map = args.value);
+      e.broadcast(Value({'name': 'alex'}));
+      expect(map['name'], equals('alex'));
     });
 
     test('Two Values as generic Event arguments', () {
@@ -134,8 +176,8 @@ void main() {
 
       var e = Event<Values<String, int>>();
       e.subscribe((args) {
-        value1 = args?.value1;
-        value2 = args?.value2;
+        value1 = args.value1;
+        value2 = args.value2;
       });
       e.broadcast(Values('boom', 37));
       expect(value1, equals('boom'));
@@ -148,8 +190,8 @@ void main() {
 
       var e = Event<Values>();
       e.subscribe((args) {
-        value1 = args?.value1;
-        value2 = args?.value2;
+        value1 = args.value1;
+        value2 = args.value2;
       });
       e.broadcast(Values('boom', 37));
       expect(value1, equals('boom'));
@@ -178,7 +220,7 @@ void main() {
       }
 
       // 3. add a handler (where applicable) to the Event
-      changeEvent.subscribe((args) => changedByEvent = args?.value);
+      changeEvent.subscribe((args) => changedByEvent = args.value);
 
       // 4. call the wrapped on... function to raise the Event
       _onChange();
@@ -222,9 +264,14 @@ void main() {
       expect(hasPassed, equals(true));
     });
 
-    test('Method toString returns the Event Type name', () {
+    test('Method toString() returns the Event name and Type when the Event is named', () {
+      var e = Event('myTestEvent');
+      expect(e.toString(), equals("myTestEvent:Event<EventArgs>"));
+    });
+
+    test('Method toString() returns UnNamed and the Type when the Event is not named', () {
       var e = Event();
-      expect(e.toString(), equals('Event<EventArgs>'));
+      expect(e.toString(), equals("UnNamed:Event<EventArgs>"));
     });
 
     // see Issue #5 - Subscribe once and exception
@@ -234,5 +281,34 @@ void main() {
       e.broadcast();
       expect(e.subscriberCount, equals(0));
     });
+
+    test('Can get the event name from args if set', () {
+      var expectedName = "myEvent";
+      String receivedName = '';
+
+      var e = Event(expectedName);
+      e.subscribe((args) => receivedName = args.eventName);
+      e.broadcast();
+
+      expect(receivedName, equals(expectedName));
+    });
+
+    test('Expect the event name to be empty if not set', () {
+      String? result; // null
+
+      var e = Event();
+      e.subscribe((args) => result = args.eventName);
+      e.broadcast();
+
+      expect(result, equals(''));
+    });
+
+    test('Broadcasting with wrong arg type should throw exception', () {
+      var e = Event<Value>();
+      e.subscribe((args) => print("Subscribed"));
+      expect(() => e.broadcast(), throwsA(TypeMatcher<ArgsError>()));
+    });
   });
+
+  group('EventArgs Tests', () {});
 }